libnet/d/bridge/i/rlkclient: move to libnet/i/rlkclient
