+++ daemon/libnetwork/drivers/bridge/bridge_linux.go
@@ -21,10 +21,10 @@ import (
 	"github.com/docker/docker/daemon/libnetwork/drivers/bridge/internal/firewaller"
 	"github.com/docker/docker/daemon/libnetwork/drivers/bridge/internal/iptabler"
 	"github.com/docker/docker/daemon/libnetwork/drivers/bridge/internal/nftabler"
-	"github.com/docker/docker/daemon/libnetwork/drivers/bridge/internal/rlkclient"
 	"github.com/docker/docker/daemon/libnetwork/drvregistry"
 	"github.com/docker/docker/daemon/libnetwork/internal/netiputil"
 	"github.com/docker/docker/daemon/libnetwork/internal/nftables"
+	"github.com/docker/docker/daemon/libnetwork/internal/rlkclient"
 	"github.com/docker/docker/daemon/libnetwork/iptables"
 	"github.com/docker/docker/daemon/libnetwork/netlabel"
 	"github.com/docker/docker/daemon/libnetwork/netutils"
